Fixes #13593 - Preload assets before sprockets
Currently Foreman core has an initializer (assets_paths) that takes care
of finding the Discovery assets in the engine and load them before
sprockets is loaded. This is happening because sprockets tries to load
the assets before they are defined (:after => finisher_hook is too
    late).

Just changing it to :before finisher_hook is enough to make it load
assets before then and the deprecation error will go away.